gatewayService/requirements.txt
blinker==1.9.0
boto3==1.35.82
botocore==1.35.82
click==8.1.7
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
pika==1.3.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
Werkzeug==3.1.3
yt-dlp==2024.12.13
